[fix] 절약 목표 응답 수정

+++ src/main/java/com/sunny/backend/dto/response/consumption/SpendTypeStatisticsResponse.java
@@ -13,6 +13,6 @@ public class SpendTypeStatisticsResponse {
     private SpendType category;
     private long totalCount;
     private long totalMoney;
-    private double percentage;
+    private long percentage;
 
 }
